Promote lodging to its own directory
