+++ client/src/pages/EventAdminLodging/index.ts
@@ -1,0 +1,4 @@
+import EventAdminLodging from './EventAdminLodging';
+// import './styles.scss';
+
+export default EventAdminLodging;
